added sign in page + some other QOL
QOL CHANGES:
made the footer be automagically put on the page... im sure this will get annoying so if theres any issues then just take it out of the layout.tsx and add it manually to the pages needed. Same thing with the gradient if it gets annoying.

SIGN IN PAGE:
Sign in page has no backend functionallity this was only templating.

//TODO: Add a button to return to home screen