Session Persistence
Session Persistence is somewhat done. The userInfo is going into the frontend but obtaining the userInfo on other pages after the redirect to /Lecturely is a bit bootstrapped. It uses the "/auth/"  which just returns the userInfo that is stored in the session store altough for a more permenant option (Because the session store is client sided and needs to be a server side thing. This is usually done through redis or dynamoStore) This means that id rather store the userinfo in a cookie in the frontend rather than getting it back every time